fixed html indentation, added more css
vim  has been giving me files with bad indentation, so I cleaned that up. Also, I had a couple more things I wanted to demo, but didnt have time for, so they can look in the file for that, such as navbar doesnt shift on link hover, letter outline writing in the header.